pyproject.toml
[tool.poetry]
name = "mcp-multi-agent-deep-researcher"
version = "0.1.0"
description = "A multi-agent research system built on the Model Context Protocol"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "Multi-Agent-deep-researcher-mcp-windows-linux"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
mcp = "^1.21.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
pydantic = "^2.5.0"
anyio = "^4.5.0"
starlette = "^0.39.0"
crewai = "^1.4.0"
crewai-tools = "^1.4.0"
litellm = "^1.79.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"